compile warnings cleaning

+++ paho_mqtt_embedded_c/MQTTClient/src/mbed/MQTTmbed.h
@@ -36,7 +36,7 @@ public:
 
     bool expired()
     {
-        return t.read_ms() >= interval_end_ms;
+        return (unsigned long)t.read_ms() >= interval_end_ms;
     }
 
     void countdown_ms(unsigned long ms)
